requirements.txt
streamlit
google-genai
python-dotenv
supabase
PyPDF2
plotly
pandas